Address organization permission review feedback
