+++ src/routes/Skills/organization-management-model.ts
@@ -9,6 +9,7 @@ import type {
 } from "../../../electron/organizations/common.ts"
 
 import { branding } from "../../../electron/branding.ts"
+export { organizationCanManage, organizationRole } from "../../lib/organization-permissions.ts"
 import { parseProviderGrants } from "./organization-provider-access.ts"
 
 export type OrganizationRole = "creator" | "member"
@@ -198,38 +199,6 @@ export function organizationNameValidation(name: string): "empty" | "invalid" | 
 
 export function allOrganizations(overview: OrganizationOverview | null): Organization[] {
   return overview ? uniqueOrganizations([...overview.created, ...overview.joined]) : []
-}
-
-export function organizationRole(
-  overview: OrganizationOverview | null,
-  organization: Organization | null,
-): OrganizationRole | null {
-  if (!overview || !organization) {
-    return null
-  }
-  if (organization.role === "creator" || organization.role === "member") {
-    return organization.role
-  }
-  if (
-    organization.creator_user_id === overview.accountId ||
-    overview.created.some((item) => item.id === organization.id)
-  ) {
-    return "creator"
-  }
-  return "member"
-}
-
-export function organizationCanManage(
-  overview: OrganizationOverview | null,
-  organization: Organization | null,
-): boolean {
-  if (!overview || !organization) {
-    return false
-  }
-  if (typeof organization.writable === "boolean") {
-    return organization.writable
-  }
-  return organizationRole(overview, organization) === "creator"
 }
 
 export function buildMemberViews(
